Early thread history experimentation. Moved some functions to a common utils file. Added more ToDos.
